refactor(soulmatch): remove compatibility wrappers and update test imports
