+++ src/screens/soulmatch/details/index.ts
@@ -1,0 +1,1 @@
+export { SoulMatchDetailsScreen } from './view/SoulMatchDetailsScreen';
