+++ src/screens/soulmatch/home/index.ts
@@ -1,0 +1,1 @@
+export { SoulMatchScreen } from './view/SoulMatchScreen';
