+++ src/screens/soulmatch/index.ts
@@ -1,0 +1,5 @@
+export { SoulMatchDetailsScreen } from './details';
+export { SoulMatchMentorScreen } from './mentor';
+export { SoulMatchRecommendationsScreen } from './recommendations';
+export { SoulMatchScreen } from './home';
+export * from './routes';
