+++ src/screens/soulmatch/mentor/index.ts
@@ -1,0 +1,1 @@
+export { SoulMatchMentorScreen } from './view/SoulMatchMentorScreen';
