+++ src/screens/soulmatch/recommendations/index.ts
@@ -1,0 +1,1 @@
+export { SoulMatchRecommendationsScreen } from './view/SoulMatchRecommendationsScreen';
